+++ src/main/java/me/pajic/misctweaks/config/ModConfig.java
@@ -15,13 +15,13 @@ public class ModConfig extends Config {
     }
 
     //? if < 1.21.5 {
-    @RequiresAction(action = Action.RESTART)
+    /*@RequiresAction(action = Action.RESTART)
     public ValidatedBoolean lodestoneChangesBackport = new ValidatedBoolean(true);
-    //?}
+    *///?}
     //? if < 1.21.6 {
-    @RequiresAction(action = Action.RESTART)
+    /*@RequiresAction(action = Action.RESTART)
     public ValidatedBoolean craftableSaddleBackport =  new ValidatedBoolean(true);
-    //?}
+    *///?}
     public ValidatedBoolean sneakingPreventsBerryBushDamage = new ValidatedBoolean(true);
     public ValidatedBoolean legArmorPreventsBerryBushDamage = new ValidatedBoolean(true);
     public ValidatedBoolean elytraSwimTweak = new ValidatedBoolean(true);