+++ src/main/java/me/pajic/misctweaks/mixin/CreeperMixin.java
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
 //? if >= 1.21.4
-/*import net.minecraft.server.level.ServerLevel;*/
+import net.minecraft.server.level.ServerLevel;
 
 @Mixin(Creeper.class)
 public abstract class CreeperMixin extends Monster {
@@ -18,22 +18,23 @@ public abstract class CreeperMixin extends Monster {
         super(entityType, level);
     }
 
+    @SuppressWarnings("resource")
     @ModifyArg(
             method = "explodeCreeper",
             at = @At(
                     value = "INVOKE",
                     //? if <= 1.21.1
-                    target = "Lnet/minecraft/world/level/Level;explode(Lnet/minecraft/world/entity/Entity;DDDFLnet/minecraft/world/level/Level$ExplosionInteraction;)Lnet/minecraft/world/level/Explosion;"
+                    /*target = "Lnet/minecraft/world/level/Level;explode(Lnet/minecraft/world/entity/Entity;DDDFLnet/minecraft/world/level/Level$ExplosionInteraction;)Lnet/minecraft/world/level/Explosion;"*/
                     //? if >= 1.21.4
-                    /*target = "Lnet/minecraft/server/level/ServerLevel;explode(Lnet/minecraft/world/entity/Entity;DDDFLnet/minecraft/world/level/Level$ExplosionInteraction;)V"*/
+                    target = "Lnet/minecraft/server/level/ServerLevel;explode(Lnet/minecraft/world/entity/Entity;DDDFLnet/minecraft/world/level/Level$ExplosionInteraction;)V"
             ),
             index = 5
     )
     private Level.ExplosionInteraction creeperExplosionDropsAll(Level.ExplosionInteraction original) {
         //? if < 1.21.4
-        boolean mobGriefing = level().getGameRules().getBoolean(GameRules.RULE_MOBGRIEFING);
+        /*boolean mobGriefing = level().getGameRules().getBoolean(GameRules.RULE_MOBGRIEFING);*/
         //? if >= 1.21.4
-        /*boolean mobGriefing = ((ServerLevel) level()).getGameRules().getBoolean(GameRules.RULE_MOBGRIEFING);*/
+        boolean mobGriefing = ((ServerLevel) level()).getGameRules().getBoolean(GameRules.RULE_MOBGRIEFING);
         return Main.CONFIG.creeperExplosionDropsAllItems.get() && mobGriefing ? Level.ExplosionInteraction.TNT : original;
     }
 }