+++ src/main/java/me/pajic/misctweaks/mixin/PlayerMixin.java
@@ -12,15 +12,15 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 //? if <= 1.21.1
-import net.minecraft.world.item.ElytraItem;
+/*import net.minecraft.world.item.ElytraItem;*/
 //? if >= 1.21.4
-/*import net.minecraft.core.component.DataComponents;*/
+import net.minecraft.core.component.DataComponents;
 
 @Mixin(Player.class)
 public abstract class PlayerMixin extends LivingEntity {
 
     //? if < 1.21.5
-    @Shadow public abstract void stopFallFlying();
+    /*@Shadow public abstract void stopFallFlying();*/
 
     protected PlayerMixin(EntityType<? extends LivingEntity> entityType, Level level) {
         super(entityType, level);
@@ -39,9 +39,9 @@ public abstract class PlayerMixin extends LivingEntity {
                 //? if < 1.21.1
                 /*isInWaterOrBubble() || isInLava() &&*/
                 //? if <= 1.21.1
-                getItemBySlot(EquipmentSlot.CHEST).getItem() instanceof ElytraItem
+                /*getItemBySlot(EquipmentSlot.CHEST).getItem() instanceof ElytraItem*/
                 //? if >= 1.21.4
-                /*getItemBySlot(EquipmentSlot.CHEST).has(DataComponents.GLIDER)*/
+                getItemBySlot(EquipmentSlot.CHEST).has(DataComponents.GLIDER)
         ) {
             stopFallFlying();
             setSwimming(true);