+++ src/main/java/me/pajic/misctweaks/mixson/ResourceModifications.java
@@ -47,7 +47,7 @@ public class ResourceModifications {
     public static void init() {
         if (FabricLoader.getInstance().isDevelopmentEnvironment()) Mixson.setDebugMode(DebugMode.EXPORT);
         //? if < 1.21.5 {
-        if (Main.CONFIG.lodestoneChangesBackport.get()) {
+        /*if (Main.CONFIG.lodestoneChangesBackport.get()) {
             Mixson.registerEvent(
                     Mixson.DEFAULT_PRIORITY,
                     "minecraft:recipe/lodestone",
@@ -105,7 +105,7 @@ public class ResourceModifications {
                             .addProperty("parent", "minecraft:adventure/root")
             );
         }
-        //?}
+        *///?}
     }
 
     public static void clientInit() {